flask_backend/requirements.txt
cohere
openai
replicate
flask
flask_cors
# transformers
python_dotenv
# torch